add covid and name pages; restructure layout with Navbar and update imports
