+++ src/app/covid/page.tsx
@@ -1,0 +1,5 @@
+const covidPage = () => {
+	return <div>Pagina covid</div>;
+};
+
+export default covidPage;
