+++ src/app/name/page.tsx
@@ -1,0 +1,9 @@
+import { NameFormLayout } from "@/layouts/NameFormLayout/NameFormLayout";
+
+const namePage = () => (
+	<>
+		<NameFormLayout />
+	</>
+);
+
+export default namePage;
