+++ src/components/NameFormComponent/hooks/useNameFormContext.ts
@@ -3,7 +3,7 @@
 // This Context hook allows our child components to easily reach into the Tabs context and get the pieces it needs.
 
 import { useContext } from "react";
-import { NameFormComponentContext } from "../context/NameFormCopmonentContext";
+import { NameFormComponentContext } from "../context/NameFormComponentContext";
 
 export const useNameFormContext = () => {
 	const context = useContext(NameFormComponentContext);
